Advanced print formatting for CLI
Modelled mostly after how Heat does it, but with some added power and
also consistency between print_list and print_dict.

We have print_list for printing of lists and print_dict for printing
of show actions. Both functions take as arguments (aside from the data
to print) also an optional dict of custom formatters (allowing to
freely format the attributes before printing) and custom labels
(allowing to tweak table headers and property names). The list is also
sorted by one of the columns.

The formatting is showcased on sample formatters and actions -
rack-show and rack-list.

Fixes bug 1213056

Change-Id: Ic14dbb930a5967e2634c1b4777e6705ab2a370ec