Revert "Support building wheels (PEP-427)"
This reverts commit f3d6ec51cf18b6d8f939f5bc0772638949ada032.

The patch being reverted should not have been merged, because
setting universal to True is used to claim in-place python 2/3
compatibility in Wheel packages.

We don't support python 3 yet (issues with keystone client), so we
must not claim the opposite here (as this would harm end users as
soon as we start uploading Wheel packages to PyPi).

Please see the following links for details:
 - http://pythonwheels.com/
 - http://lists.openstack.org/pipermail/openstack-dev/2013-November/020059.html

Change-Id: I71b73998fe22aaf11cb39d26b428cbfe567c0398