[WIP] Update tuskarclient to match API changes
Updated the tuskarclient to reflect the changes made in the Tuskar API.
This consists of removing Racks, Flavors and...

This patch is very much a work in progress, it is dependant on the
changes in the Tuskar API and also terminology discussions in the
mailing list. The initial work is to remove some of the older concepts,
and start to add in the new ones.

So, quite a bit away from being ready.

Implements: blueprint update-to-match-api-changes
Change-Id: I5dc2ad40bc9a61080084db229859a7875635e2b6