Remove multivalue attributes and roles from CLI
This change removes the option of having multiple values on the CLI.
Previously `--attributes key=1;key=2` was valid but these now need to be
provided individually as `--attribute key=1 --attribute key=2`. To make
this slightly less verbose a short alias has been added and the
following can be used `-A key=1 -A key=2`

This change applies to both --attributes and --roles.

Fixes bug 1284631

Change-Id: I4d39310d702d41f8231b14933d4db36f40f49366