Revert "Use HTTPClient from common apiclient code"
This reverts commit e83c56584ea182255af0ae54cf71104a77c6de1b.

This code was failing to pass along the correct headers when trying to
retrieve the Tuskar endpoint from Keystone. However, it worked upon
reverting this commit.

The reversion exposed an error in the tests for python 3, which is fixed
as described in this abandoned patch: https://review.openstack.org/#/c/65673/

Closes-Bug: #1307945
Change-Id: I17fa9ddd8bd1722f207393bb1249249c81c1342b