Reimplement print_dict and print_list formatters
This partially reverts 948e05d which removed the formatters.
The removal is incorrect as the one in oslo doesn't have the same
functionality. Excessive mocking in the tests hid this problem
originally, so this change adds a test to avoid this in the future.

The plan is to remove these again with a larger refactor, but at the
moment they break master.

Change-Id: I1d075dca66cbaa5d2ae8c4e5034373fb392ff5d9